# runtime-platform/contracts/openapi/control.v1.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "VitalServer Runtime Platform Control API",
    "version": "1.0.0",
    "description": "Public control-plane API v1. Every read returns an owner-supplied ReadResult. Commands return a durable Operation, a CommandRejection when they were not admitted, or a typed admission/forwarding failure when existence cannot safely be inferred; they do not synthesize state from another owner."
  },
  "paths": {
...
    "/v1/runtime/topology:apply": {
      "post": {
        "operationId": "applyRuntimeTopology",
        "summary": "Request topology application",
        "x-owner": "guest-runtime",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "../json-schema/v1/runtime-topology-apply-command.schema.json"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "The command was accepted and has a durable operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../json-schema/v1/operation.schema.json"
                }
              }
            }
          },
          "400": {
            "description": "The input was rejected before an operation existed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../json-schema/v1/command-rejection.schema.json"
                }
              }
            }
          },
          "500": {
            "description": "The Guest Runtime could not determine command admission. Retry the exact requestId after Guest recovery.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandAdmissionFailure"
                }
              }
            }
          },
          "503": {
            "description": "The Host facade could not reach the configured Guest endpoint before command admission; no Guest operation exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../json-schema/v1/command-rejection.schema.json"
                }
              }
            }
          },
          "502": {
            "description": "The Host facade attempted command forwarding but cannot state whether the Guest command exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../json-schema/v1/facade-forwarding-failure.schema.json"
                }
              }
            }
          }
        }
      }
    },
    "/v1/runtime/operations/{operationId}": {
      "get": {
        "operationId": "getRuntimeOperation",
        "summary": "Read a durable Runtime operation",
        "x-owner": "guest-runtime",
        "parameters": [
          {
            "$ref": "#/components/parameters/OperationId"
          }
        ],
        "responses": {
          "200": {
            "description": "Explicit operation read result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationReadResult"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/runtime/lab/sessions": {
      "get": {
        "operationId": "listLabSessions",
        "summary": "List Guest Runtime Lab-owned sessions",
        "x-owner": "guest-runtime-lab",
        "responses": {
          "200": {
            "description": "Explicit Lab session collection read result",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/LabSessionCollectionReadResult" } } }
          }
        }
      },
      "post": {
        "operationId": "createLabSession",
        "summary": "Create a prepared Lab session and its owned beds and virtual recorders",
        "x-owner": "guest-runtime-lab",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "../json-schema/v1/lab-commands.schema.json#/$defs/CreateLabSessionCommand" } } }
        },
        "responses": {
          "202": { "description": "Created durable Lab operation", "content": { "application/json": { "schema": { "$ref": "../json-schema/v1/operation.schema.json" } } } },
          "400": { "description": "The command was rejected before a Lab operation existed", "content": { "application/json": { "schema": { "$ref": "../json-schema/v1/command-rejection.schema.json" } } } },
          "500": { "description": "Lab command admission outcome is unknown", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CommandAdmissionFailure" } } } }
        }
      }
    },
    "/v1/runtime/lab/sessions/{sessionId}": {
      "get": {
        "operationId": "getLabSession",
        "summary": "Read one Guest Runtime Lab-owned session",
        "x-owner": "guest-runtime-lab",
        "parameters": [{ "$ref": "#/components/parameters/SessionId" }],
        "responses": {
          "200": { "description": "Explicit Lab session read result", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/LabSessionReadResult" } } } }
        }
      }
    },
...
    "/v1/runtime/lab/beds/{bedId}": {
      "get": {
        "operationId": "getLabBed",
        "summary": "Read one Guest Runtime Lab-owned bed",
        "x-owner": "guest-runtime-lab",
        "parameters": [{ "$ref": "#/components/parameters/BedId" }],
        "responses": {
          "200": { "description": "Explicit Lab bed read result", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/LabBedReadResult" } } } }
        }
      }
    },
...
    "/v1/runtime/lab/recorders/{virtualRecorderId}": {
      "get": {
        "operationId": "getVirtualRecorder",
        "summary": "Read one Guest Runtime Lab-owned virtual recorder",
        "x-owner": "guest-runtime-lab",
        "parameters": [{ "$ref": "#/components/parameters/VirtualRecorderId" }],
        "responses": {
          "200": { "description": "Explicit virtual recorder read result", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/VirtualRecorderReadResult" } } } }
        }
      }
    },
    "/v1/runtime/lab/resources:command": {
      "post": {
        "operationId": "executeLabResourceCommand",
        "summary": "Execute an explicit Lab start, stop, visibility, detach, or delete command",
        "x-owner": "guest-runtime-lab",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "../json-schema/v1/lab-commands.schema.json#/$defs/LabResourceCommand" } } }
        },
        "responses": {
          "202": { "description": "Created durable Lab operation", "content": { "application/json": { "schema": { "$ref": "../json-schema/v1/operation.schema.json" } } } },
          "400": { "description": "The command was rejected before a Lab operation existed", "content": { "application/json": { "schema": { "$ref": "../json-schema/v1/command-rejection.schema.json" } } } },
          "500": { "description": "Lab command admission outcome is unknown", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CommandAdmissionFailure" } } } }
        }
      }
    },
    "/v1/runtime/lab/deletion-receipts/{receiptId}": {
      "get": {
        "operationId": "getLabDeletionReceipt",
        "summary": "Read explicit Lab deletion and retained-artifact evidence",
        "x-owner": "guest-runtime-lab",
        "parameters": [{ "$ref": "#/components/parameters/ReceiptId" }],
        "responses": {
          "200": { "description": "Explicit deletion receipt read result", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/DeletionReceiptReadResult" } } } }
        }
      }
    },
    "/v1/runtime/archive/exports": {
      "post": {
        "operationId": "exportLabArtifact",
        "summary": "Finalize a stopped virtual recorder source and export its artifact",
        "x-owner": "archive-export",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "../json-schema/v1/artifact-export-command.schema.json" } } }
        },
        "responses": {
          "202": { "description": "Created durable Archive Export operation", "content": { "application/json": { "schema": { "$ref": "../json-schema/v1/operation.schema.json" } } } },
          "400": { "description": "The command was rejected before an Archive operation existed", "content": { "application/json": { "schema": { "$ref": "../json-schema/v1/command-rejection.schema.json" } } } },
          "500": { "description": "Archive command admission outcome is unknown", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CommandAdmissionFailure" } } } }
        }
      }
    },
    "/v1/runtime/archive/manifests/{artifactId}": {
      "get": {
        "operationId": "getRuntimeArtifactManifest",
        "summary": "Read a Guest Runtime Archive Export immutable manifest",
        "x-owner": "archive-export",
        "parameters": [{ "$ref": "#/components/parameters/ArtifactId" }],
        "responses": {
          "200": { "description": "Explicit artifact manifest read result", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ArtifactManifestReadResult" } } } }
        }
      }
    },
    "/v1/runtime/archive/export-receipts/{receiptId}": {
      "get": {
        "operationId": "getRuntimeExportReceipt",
        "summary": "Read a Guest Runtime Archive Export receipt",
        "x-owner": "archive-export",
        "parameters": [{ "$ref": "#/components/parameters/ReceiptId" }],
        "responses": {
          "200": { "description": "Explicit Archive Export receipt read result", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ExportReceiptReadResult" } } } }
        }
      }
    },
    "/v1/runtime/external-upstreams": {
      "get": {
        "operationId": "listExternalUpstreamIntegrations",
        "summary": "List Guest Runtime-owned external upstream integrations",
        "x-owner": "guest-runtime-external-upstream",
        "responses": {
          "200": { "description": "Explicit external integration collection read", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ExternalUpstreamIntegrationCollectionReadResult" } } } }
        }
      },
      "post": {
        "operationId": "applyExternalUpstreamIntegration",
        "summary": "Persist and explicitly observe an external upstream integration",
        "x-owner": "guest-runtime-external-upstream",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "../json-schema/v1/external-upstream-command.schema.json" } } } },
        "responses": {
          "202": { "description": "Durable external integration operation", "content": { "application/json": { "schema": { "$ref": "../json-schema/v1/operation.schema.json" } } } },
          "400": { "description": "Command rejected before admission", "content": { "application/json": { "schema": { "$ref": "../json-schema/v1/command-rejection.schema.json" } } } },
          "500": { "description": "External integration admission outcome is unknown", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CommandAdmissionFailure" } } } }
        }
      }
    },
    "/v1/runtime/external-upstreams/{integrationId}": {
      "get": {
        "operationId": "getExternalUpstreamIntegration",
        "summary": "Read one Guest Runtime-owned external upstream integration",
        "x-owner": "guest-runtime-external-upstream",
        "parameters": [{ "$ref": "#/components/parameters/IntegrationId" }],
        "responses": {
          "200": { "description": "Explicit external integration read", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ExternalUpstreamIntegrationReadResult" } } } }
        }
      }
    },
...
    "/v1/runtime/relay-targets/{targetId}": {
      "get": {
        "operationId": "getOutboundRelayTarget",
        "summary": "Read one Guest Runtime-owned outbound relay target",
        "x-owner": "guest-runtime-relay",
        "parameters": [{ "$ref": "#/components/parameters/RelayTargetId" }],
        "responses": {
          "200": { "description": "Explicit relay target read", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/OutboundRelayTargetReadResult" } } } }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
      "OperationId": {
        "name": "operationId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        }
      },
...
      "SessionId": {
        "name": "sessionId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        }
      },
      "BedId": {
        "name": "bedId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        }
      },
...
    },
    "schemas": {
...
      "CommandAdmissionFailure": {
        "$ref": "../json-schema/v1/command-admission-failure.schema.json"
      },
      "OperationReadResult": {
        "allOf": [
          {
            "$ref": "../json-schema/v1/read-result.schema.json"
          },
          {
            "type": "object",
            "properties": {
              "value": {
                "$ref": "../json-schema/v1/operation.schema.json"
              }
            }
          }
        ]
      },
...
      "LabSessionCollectionReadResult": {
        "allOf": [
          { "$ref": "../json-schema/v1/read-result.schema.json" },
          {
            "type": "object",
            "properties": {
              "value": {
                "type": "array",
                "items": { "$ref": "../json-schema/v1/lab-resources.schema.json#/$defs/LabSession" }
              }
            }
          }
        ]
      },
...
    }
  }
}